backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
asyncpg
redis
celery
anthropic
crewai
python-dotenv
pydantic-settings
python-jose
passlib==1.7.4
bcrypt==4.3.0
websockets
slowapi
httpx
cryptography
